# Conflicts:
#	app/src/main/java/com/xjek/provider/views/dashboard/DashBoardViewModel.kt

+++ app/src/main/java/com/xjek/provider/views/dashboard/DashBoardViewModel.kt
@@ -22,7 +22,8 @@ class DashBoardViewModel : BaseViewModel<DashBoardNavigator>() {
     var currentStatus = MutableLiveData<String>()
 
     fun callCheckStatusAPI() {
-        if (latitude.value!! > 0 && longitude.value!! > 0) {
+//        if (latitude.value!! > 0 && longitude.value!! > 0) {
+        if (latitude.value!!.toInt() > 0 && longitude.value!!.toInt() > 0) {
             if (readPreferences<String>(PreferencesKey.ACCESS_TOKEN).length > 2)
                 getCompositeDisposable().add(appRepository.checkRequest(this,
                         "Bearer " + readPreferences<String>(PreferencesKey.ACCESS_TOKEN),
